+++ theme/animations/AddingStopLossAnimation.tsx
@@ -1,6 +1,6 @@
 import Lottie from 'react-lottie'
 import { Box } from 'theme-ui'
-import addingStopLossAnimation from 'theme/lottie/addingStopLossAnimation.json'
+import addingStopLossAnimation from 'theme/animations/lottie/addingStopLossAnimation.json'
 
 const addingStopLossAnimationOptions = {
   loop: true,
