+++ theme/animations/lottie/addingStopLossAnimation.json
