+++ theme/animations/lottie/openVaultAnimation.json
